change socialblock structure, separate data-engine from backend
